rust/updateToolchain: Use `github-tags` datasource
Patch releases often only get a tag in `rust-lang/rust` but not a proper "release". This leads to those patch releases not showing up as renovatebot PRs.

This change is an attempt to make them show up regardless of only being a tag without attached release notes. Ideally, renovatebot is clever enough to find the release notes in the `RELEASES.md` file, but we'll see how well this will work in practice.